- moved system_has_windows? to Timezone module
