+++ package/yast2-country.spec
@@ -37,6 +37,7 @@ BuildRequires:  rubygem(rspec)
 BuildRequires:  yast2-core >= 3.1.12
 # RSpec extensions for YaST
 BuildRequires:  yast2-ruby-bindings >= 3.1.26
+BuildRequires:  yast2-storage-ng
 
 Requires:	yast2-trans-stats yast2-perl-bindings timezone
 # XVersion.ycp
